Python ranking module available for 2.7 and 3

+++ pythonRankingModule/main_vIntgrt.py
@@ -11,7 +11,7 @@
 # Copyright:     (c) ITI (CERTH) 2015
 # Licence:       <apache licence 2.0>
 #-------------------------------------------------------------------------------
-import time, os, sys, socket
+import time, os, sys, socket, glob
 '''Check for dependencies'''
 try:
     import igraph
@@ -21,6 +21,7 @@ except:
 from CommunityRanking_vIntgrt import communityranking
 #-------------------------------
 print(time.asctime( time.localtime(time.time()) ))
+t = time.time()
 
 '''PARAMETERS'''
 # User sets mongo host
@@ -34,73 +35,78 @@ except:
 try:
     dataCollection = sys.argv[2]
 except:
-    dataCollection = 'snowDataset'
+    dataCollection = 'testDataset'
     pass
-
+# User sets timestamp_start
 try:
     lowerTime = int(sys.argv[3])
 except:
     lowerTime = False#(1393342340+3600*10)*1000#False#
     pass
+# User sets timestamp_end
 try:
     upperTime = int(sys.argv[4])
 except:
     upperTime = False#(1393431754-3600*10)*1000#False#
     pass
-
+# User sets json writing path for visualization module
 try:
     jsonWritingPath = (sys.argv[5])
 except:
     jsonWritingPath = os.getcwd()
 
-if not os.path.exists('./tmp/'):
-    os.makedirs('./tmp/')
+if not os.path.exists(jsonWritingPath+'/Com_Graph/jsons/'+dataCollection+'communities.json'):
 
-print(dataCollection)
+    if not os.path.exists('./tmp/'):
+        os.makedirs('./tmp/')
 
-t = time.time()
+    print(dataCollection)
 
-'''Functions'''
+    '''Functions'''
 
-data = communityranking.from_json(mongoHost, dataCollection, lowerTime, upperTime)
-elapsed = time.time() - t
-print('Stage 1: %.2f seconds' % elapsed)
+    data = communityranking.from_json(mongoHost, dataCollection, lowerTime, upperTime)
+    elapsed = time.time() - t
+    print('Stage 1: %.2f seconds' % elapsed)
 
-#User sets how many timeslots back the framework should search
-prevTimeslots = 3
-dataEvol=data.evol_detect(prevTimeslots)
-del(data)
-elapsed = time.time() - t
-print('Stage 3: %.2f seconds' % elapsed)
+    #User sets how many timeslots back the framework should search
+    prevTimeslots = 3
+    dataEvol=data.evol_detect(prevTimeslots)
+    del(data)
+    elapsed = time.time() - t
+    print('Stage 3: %.2f seconds' % elapsed)
 
-print("Ranking Commences")
-numTopComms = 20 #how many dynamic communities to create illustrations for
-rankedCommunities = dataEvol.commRanking(numTopComms,jsonWritingPath)
+    print("Ranking Commences")
+    numTopComms = 20 #how many dynamic communities to create illustrations for
+    rankedCommunities = dataEvol.commRanking(numTopComms,jsonWritingPath)
 
-os.remove('./tmp/'+dataCollection+'UserDict.pck')
+    os.remove('./tmp/'+dataCollection+'UserDict.pck')
 
-#send success message to rabbitMQ server
-try:
-    import pika
-    credentials = pika.PlainCredentials('test', 'test')
+    jsonFiles = glob.glob('./tmp/*.json')
+
+else:
+
+    #send success message to rabbitMQ server
     try:
-        connection = pika.BlockingConnection(pika.ConnectionParameters('localhost'))
+        import pika
+        credentials = pika.PlainCredentials('test', 'test')
+        try:
+            connection = pika.BlockingConnection(pika.ConnectionParameters('localhost'))
+        except:
+            connection = pika.BlockingConnection(pika.ConnectionParameters(host='160.40.50.236', credentials=credentials))
+            pass
+        channel = connection.channel()
+        channel.queue_declare(queue='success')
+        channel.basic_publish(exchange='',routing_key='success',body='SUCCESS')
+        connection.close()
     except:
-        connection = pika.BlockingConnection(pika.ConnectionParameters(host='160.40.50.236', credentials=credentials))
         pass
-    channel = connection.channel()
-    channel.queue_declare(queue='success')
-    channel.basic_publish(exchange='',routing_key='success',body='SUCCESS')
-    connection.close()
-except:
-    pass
 
-import webbrowser
-try:
-    webbrowser.get('firefox').open_new_tab('file:///'+jsonWritingPath+'/Com_Graph/community.html?collection='+dataCollection)
-except:
-    getFirefox = webbrowser.get("c:/program files (x86)/mozilla firefox/firefox.exe %s &")
-    getFirefox.open_new_tab('file:///'+jsonWritingPath+'/Com_Graph/community.html?collection='+dataCollection)
+    import webbrowser
+    try:
+        webbrowser.get('firefox').open_new_tab('file:///'+jsonWritingPath+'/Com_Graph/community.html?collection='+dataCollection)
+    except:
+        getFirefox = webbrowser.get("c:/program files (x86)/mozilla firefox/firefox.exe %s &")
+        getFirefox.open_new_tab('file:///'+jsonWritingPath+'/Com_Graph/community.html?collection='+dataCollection)
 
 elapsed = time.time() - t
 print('Elapsed: %.2f seconds' % elapsed)
