+++ pythonRankingModule/main_vIntgrt_27.py
@@ -32,14 +32,15 @@ except:
 try:
     dataCollection = sys.argv[2]
 except:
-    dataCollection = 'snowDataset'
+    dataCollection = 'testDataset'
     pass
-
+# User sets timestamp_start
 try:
     lowerTime = int(sys.argv[3])
 except:
     lowerTime = False
     pass
+# User sets timestamp_end
 try:
     upperTime = int(sys.argv[4])
 except:
@@ -50,49 +51,52 @@ try:
     jsonWritingPath = (sys.argv[5])
 except:
     jsonWritingPath = os.getcwd()
-    
-if not os.path.exists('./tmp/'):
-    os.makedirs('./tmp/')    
 
-print dataCollection
+if not os.path.exists(jsonWritingPath+'/Com_Graph/jsons/'+dataCollection+'communities.json'):
 
-t = time.time()
+    if not os.path.exists('./tmp/'):
+        os.makedirs('./tmp/')    
 
-'''Functions'''
+    print dataCollection
 
-data = communityranking.from_json(mongoHost, dataCollection, lowerTime, upperTime)
-elapsed = time.time() - t
-print 'Stage 1: %.2f seconds' % elapsed
+    t = time.time()
 
-#User sets how many timeslots back the framework should search
-prevTimeslots = 3
-dataEvol=data.evol_detect(prevTimeslots)
-del(data)
-elapsed = time.time() - t
-print 'Stage 3: %.2f seconds' % elapsed
+    '''Functions'''
 
-print u"Ranking Commences"
-numTopComms = 20 #how many dynamic communities to create illustrations for
-rankedCommunities = dataEvol.commRanking(numTopComms,jsonWritingPath)
+    data = communityranking.from_json(mongoHost, dataCollection, lowerTime, upperTime)
+    elapsed = time.time() - t
+    print 'Stage 1: %.2f seconds' % elapsed
 
-os.remove('./tmp/'+dataCollection+'UserDict.pck')
+    #User sets how many timeslots back the framework should search
+    prevTimeslots = 3
+    dataEvol=data.evol_detect(prevTimeslots)
+    del(data)
+    elapsed = time.time() - t
+    print 'Stage 3: %.2f seconds' % elapsed
+
+    print u"Ranking Commences"
+    numTopComms = 20 #how many dynamic communities to create illustrations for
+    rankedCommunities = dataEvol.commRanking(numTopComms,jsonWritingPath)
+
+    os.remove('./tmp/'+dataCollection+'UserDict.pck')
+
+else:
+
+    #send success message to rabbitMQ server
+    try:
+        import pika
+        credentials = pika.PlainCredentials('test', 'test')
+        try:
+            connection = pika.BlockingConnection(pika.ConnectionParameters('localhost'))
+        except:
+            connection = pika.BlockingConnection(pika.ConnectionParameters(host='160.40.50.236', credentials=credentials))
+            pass
+        channel = connection.channel()
+        channel.queue_declare(queue='success')
+        channel.basic_publish(exchange='',routing_key='success',body='SUCCESS')
+        connection.close()
+    except:
+        pass
 
 elapsed = time.time() - t
 print 'Elapsed: %.2f seconds' % elapsed
-#send success message to rabbitMQ server
-try:
-    import pika
-    credentials = pika.PlainCredentials('test', 'test')
-    try:
-        connection = pika.BlockingConnection(pika.ConnectionParameters('localhost'))
-    except:
-        connection = pika.BlockingConnection(pika.ConnectionParameters(host='160.40.50.236', credentials=credentials))
-        pass
-    channel = connection.channel()
-    channel.queue_declare(queue='success')
-    channel.basic_publish(exchange='',routing_key='success',body='SUCCESS')
-    connection.close()
-except:
-    pass
-
-
